Move base mixin into its own file
This fixes the too long file pylint warning and encompasses common
functionality together since ChangeTracked, ContextRBAC, Identifiable
mixins are just used in Base and nowhere else.

This commit does not contain and code changes, only cut paste from one
file to the other. Additional changes will be in separate commits.
